+++ src/main/resources/application.properties
@@ -1,13 +1,11 @@
 logging.file.name=/var/log/spring.log
-spring.profiles.active=international, exploratory, fast-polyline
 #---
 spring.config.activate.on-profile=exploratory
 WAYPOINT_GENERATOR_AREA_DIVISION = 128
 WAYPOINT_GENERATOR_NUMBER_OF_RESULTS = 5
 DISTANCE_FACTOR = 1.0
-#---
-spring.config.activate.on-profile=international
 GOOGLE_API_KEY=${GOOGLE_API_KEY}
+TMAP_API_KEY=${TMAP_API_KEY}
 #---
 spring.config.activate.on-profile=fast-polyline
 REQUEST_FIELDMASK = routes.duration,routes.distanceMeters,routes.polyline.encodedPolyline
@@ -17,5 +15,3 @@ spring.config.activate.on-profile=detail-route
 REQUEST_FIELDMASK = routes.duration,routes.distanceMeters,routes.legs
 GET_POLYLINE_DIRECTLY = false
 #---
-spring.config.activate.on-profile=korea
-TMAP_API_KEY=${TMAP_API_KEY}
